requirements.txt
langchain
langgraph
langchain_community
pandas
python-dotenv
langchain-google-genai
pydantic
langchain[google-genai]
alpha_vantage
tavily-python